:recycle: Refactor LOCALE constants
