+++ src/components/common/sidebar/LanguageChanger.tsx
@@ -4,9 +4,9 @@ import { useTranslation } from 'react-i18next';
 import { usePathname, useRouter } from 'next/navigation';
 import { i18nConfig } from '@/i18n/i18nConfig';
 import React from 'react';
-import LOCALE from '../../../../lib/constants/LOCALE';
+import { INFOS } from '../../../../lib/constants/LOCALE';
 
-const localeInfo = Object.entries(LOCALE.INFOS).map(([id, info]) => {
+const localeInfo = Object.entries(INFOS).map(([id, info]) => {
   return { id: id, text: info.buttonText };
 });
