+++ src/components/common/sidebar/SideBar.tsx
@@ -5,7 +5,7 @@ import MenuHeader from '@/components/common/sidebar/MenuHeader';
 import Link from 'next/link';
 import { useTranslation } from 'react-i18next';
 import useMenuOpen from '@/hooks/useMenuOpen';
-import { SongsBySinger, SongType } from '../../../../types/songs';
+import { SongsBySinger, SongType } from '../../../../types/song';
 
 interface Props {
   songsBySingerList: SongsBySinger[];
@@ -27,13 +27,13 @@ export default function SideBar({ songsBySingerList }: Props) {
           <div key={singer}>
             {singer}
             {songs.map(({ title, slug }: Pick<SongType, 'title' | 'slug'>) => (
-              <p key={title}>
-                {title} {slug}
-              </p>
+              <Link href={slug} key={title} onClick={closeMenu}>
+                <p>{title}</p>
+              </Link>
             ))}
           </div>
         ))}
-        <Link className="text-blue-600" href="/about-us">
+        <Link className="text-blue-600" href="/about-us" onClick={closeMenu}>
           {t('common:about_us')}
         </Link>
       </aside>
